requirements.txt
requests
PyYAML
psycopg2
multiprocessing
